+++ WinRTTextToSpeech/UnityAddon/WinRTTextToSpeechPinvokes.cs
@@ -18,6 +18,16 @@ namespace Microsoft.MixedReality.Toolkit.Speech.Windows
             CallingConvention = CallingConvention.StdCall)]
         public static extern bool TrySynthesizePhrase(
             string phrase,
+            out IntPtr data,
+            out int length);
+
+        [DllImport(
+            "WinRTTextToSpeech.dll",
+            EntryPoint = "TrySynthesizePhraseWithCustomVoice",
+            CharSet = CharSet.Ansi,
+            CallingConvention = CallingConvention.StdCall)]
+        public static extern bool TrySynthesizePhraseWithCustomVoice(
+            string phrase,
             string voiceName,
             out IntPtr data,
             out int length);